added timeline to projects, moved connect modal to navbar, added project entry
