+++ src/app/footer/footer.component.css
@@ -1,11 +1,2 @@
-.contact-btn {
-  background-color: #afc47d; /* Green */
-  border: none;
-  border-radius: 5px;
-  color: white !important;
-  padding: 3px 10px;
-  text-align: center;
-  text-decoration: none;
-  display: inline-block;
-}
+
