+++ src/app/footer/footer.component.html
@@ -1,5 +1,5 @@
 <footer class="border-top footer text-muted">
   <div class="container">
-    2023 - Chayanne Rodriguez - <button class="contact-btn" color="#176177" (click)="open()">Contact</button>
+    2023 <i class="bi bi-flower3"></i> Chayanne Rodriguez
   </div>
 </footer>
