+++ src/app/navbar/navbar.component.html
@@ -13,6 +13,9 @@
           <li class="nav-item">
             <a routerLink="projects" class="nav-link text-dark fancy-txt">P R O J E C T S</a>
           </li>
+          <li class="nav-item">
+            <a [routerLink]="[]" (click)="open()" class="nav-link text-dark fancy-txt">C O N N E C T</a>
+          </li>
         </ul>
       </div>
     </div>
