+++ src/app/projects/projects.component.css
@@ -23,10 +23,29 @@ h1 {
 .projects {
   display: flex;
   flex-direction: column;
-  margin: 30px 30px 30px;
+  margin: 30px 30px 80px;
   padding: 20px;
+  gap: 20px;
   border-radius: 9px;
   background-color: #f6f2ec;
+}
+
+.timeline-line {
+  display: none;
+}
+
+.project-item {
+  display: flex;
+  flex-direction: column;
+}
+.project-date {
+  color: #afc47d;
+  font-family: playfair-display, serif;
+  font-size: 15px;
+}
+
+.project-description {
+  margin: 3px 10px 10px;
 }
 
 li::marker {
@@ -35,25 +54,18 @@ li::marker {
   font-weight: bolder;
 }
 
-.text {
-  color: #f2f2f2;
-  font-size: 14px;
-  padding: 8px 12px;
-  position: absolute;
-  bottom: 8px;
-  width: 100%;
-  text-align: center;
-}
-
 .web-gif {
   width: 250px;
+  display: block;
   margin: auto;
+  text-align: center;
+  border-radius: 9px;
 }
 
 ngb-carousel {
   margin: auto;
   max-width: 600px;
-  text-align: center;
+  text-align: center !important;
 }
 
 .youtube-player {
@@ -73,16 +85,6 @@ ngb-carousel {
   left: 0;
   width: 100%;
   height: 100%;
-}
-
-.btn-time {
-  background: #afc47d;
-  color: #f6f2ec;
-  border-color: #afc47d;
-  border-radius: 8px;
-  border-style: solid;
-  padding: 5px;
-  margin-right: 5px;
 }
 
 .btn-code {
@@ -125,7 +127,25 @@ ngb-carousel {
   .projects-intro {
      margin: 30px 20%;
   }
+  .project-item {
+    flex-direction: row;
+    gap: 10px;
+  }
 
+  .timeline {
+    width: 140px;
+    text-align: center;
+  }
+  .timeline-line {
+    display: block;
+    border-left: 2px solid #afc47d;
+    position: relative;
+    left: 50%;
+    height: 98%;
+  }
+  .project-description {
+    flex: 1;
+  }
   .web-gif {
     width: 500px;
   }
@@ -136,6 +156,23 @@ ngb-carousel {
     margin: 30px 30%;
   }
   .projects {
-    margin: 30px 200px 30px;
+    margin: 30px 100px 80px;
+  }
+  .timeline {
+    width: 150px;
+  }
+  .project-date {
+    font-size: medium;
+  }
+}
+
+@media screen and (min-width: 1920px) {
+  .projects {
+    margin: 30px 200px 80px;
+    padding-left: 80px;
+    padding-right: 80px;
+  }
+  .project-item {
+    gap: 30px;
   }
 }
