+++ src/styles.css
@@ -1,10 +1,10 @@
  /* You can add global styles to this file, and also import other style files */
 @import url('https://use.typekit.net/gsr7uey.css');
 @import url('https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css');
+@import url('https://cdn.jsdelivr.net/npm/bootstrap-icons@1.10.5/font/bootstrap-icons.css');
 
 html {
   font-size: 14px;
-  /*position: relative;*/
   min-height: 100vh;
 }
 
@@ -15,6 +15,14 @@ body {
   position: relative;
 }
 
+footer {
+ background-color: #f6f2ec;
+ position: absolute;
+ bottom: 0;
+ padding: 10px;
+ width: 100%;
+}
+
 .fancy-txt {
  font-family: playfair-display, serif;
  font-weight: 700 !important;
@@ -22,7 +30,7 @@ body {
 }
 
 a:link, a:visited {
-  color: #176177;
+  color: #84964d;
 }
 
 @media (min-width: 768px) {
@@ -31,18 +39,7 @@ a:link, a:visited {
   }
 }
 
-body {
-  margin-bottom: 60px;
-}
-
-footer {
- background-color: #f6f2ec;
- position: absolute;
-  bottom: 0;
-  padding: 10px;
-  width: 100%;
-}
-
+ /* loading screen ellipsis animation*/
 .lds-ellipsis {
  display: inline-block;
  position: relative;
